refactor(pipelines): extract daily pipeline to lib/pipelines/daily.js; delegate from index.js
